+++ ccdproc/tests/test_image_collection.py
@@ -160,6 +160,53 @@ class TestImageFileCollection(object):
         new_data = np.array(collection.summary)
         assert (new_data == old_data).all()
 
+    @pytest.mark.parametrize('extension', ['TESTEXT', 1, ('TESTEXT', 1)])
+    def test_multiple_extensions(self, triage_setup, extension):
+        ext1 = fits.PrimaryHDU()
+        ext1.data = np.arange(1, 5)
+        # It is important than the name used for this test extension
+        # NOT be MASK or UNCERT because both are treated in a special
+        # way by the FITS reader.
+        test_ext_name = 'TESTEXT'
+        ext2 = fits.ImageHDU(name=test_ext_name)
+        ext2.data = np.arange(6, 10)
+        hdulist = fits.hdu.hdulist.HDUList([ext1, ext2])
+
+        hdulist.writeto(os.path.join(triage_setup.test_dir,
+                                     'multi-extension.fits'))
+        ic2 = image_collection.ImageFileCollection(
+            triage_setup.test_dir, keywords='*',
+            filenames=['multi-extension.fits'], ext=extension)
+
+        ic1 = image_collection.ImageFileCollection(triage_setup.test_dir,
+                    keywords='*', filenames=['multi-extension.fits'], ext=0)
+
+        assert ic1.ext == 0
+        assert ic2.ext == extension
+
+        column2 = ic2.summary_info.colnames
+        column1 = ic1.summary_info.colnames
+
+        assert column1 != column2
+
+        list1 = [key.lower() for key in ext2.header]
+        list2 = ic2.summary_info.colnames[1:]
+
+        assert list1 == list2
+
+        ccd_kwargs = {'unit': 'adu'}
+        for data, hdr, hdu, ccd in zip(ic2.data(),
+                                       ic2.headers(),
+                                       ic2.hdus(),
+                                       ic2.ccds(ccd_kwargs)):
+            np.testing.assert_array_equal(data, ext2.data)
+            assert hdr == ext2.header
+            # Now compare that the generators each give the same stuff
+            np.testing.assert_array_equal(data, ccd.data)
+            np.testing.assert_array_equal(data, hdu.data)
+            assert hdr == hdu.header
+            assert hdr == ccd.meta
+
     def test_headers(self, triage_setup):
         collection = image_collection.ImageFileCollection(location=triage_setup.test_dir,
                                              keywords=['imagetyp'])